+++ core/menu_stack.js
@@ -5,12 +5,12 @@
 const loadMenu = require('./menu_util.js').loadMenu;
 const { Errors, ErrorReasons } = require('./enig_error.js');
 const { getResolvedSpec } = require('./menu_util.js');
+const { MenuFlags } = require('./menu_module.js');
 
 //  deps
 const _ = require('lodash');
 const assert = require('assert');
-
-//  :TODO: Stack is backwards.... top should be most recent! :)
+const bunyan = require('bunyan');
 
 module.exports = class MenuStack {
     constructor(client) {
@@ -27,19 +27,11 @@ module.exports = class MenuStack {
     }
 
     peekPrev() {
-        if (this.stackSize > 1) {
-            return this.stack[this.stack.length - 2];
-        }
+        return this.stack[this.stack.length - 2];
     }
 
     top() {
-        if (this.stackSize > 0) {
-            return this.stack[this.stack.length - 1];
-        }
-    }
-
-    get stackSize() {
-        return this.stack.length;
+        return this.stack[this.stack.length - 1];
     }
 
     get currentModule() {
@@ -81,19 +73,15 @@ module.exports = class MenuStack {
     prev(cb) {
         const menuResult = this.top().instance.getMenuResult();
 
-        const currentModuleInfo = this.top();
-
-        //  :TODO: leave() should really take a cb...
         this.pop().instance.leave(); //  leave & remove current
 
-        const previousModuleInfo = this.pop(); //  get previous
+        const previousModuleInfo = this.pop(); //  get previous; we'll re-create a instance
 
         if (previousModuleInfo) {
             const opts = {
                 extraArgs: previousModuleInfo.extraArgs,
                 savedState: previousModuleInfo.savedState,
                 lastMenuResult: menuResult,
-                currentModuleInfo,
             };
 
             return this.goto(previousModuleInfo.name, opts, cb);
@@ -111,8 +99,7 @@ module.exports = class MenuStack {
         }
         options = options || {};
 
-        const currentModuleInfo = options.currentModuleInfo || this.top();
-        const self = this;
+        const currentModuleInfo = this.top();
 
         if (currentModuleInfo && name === currentModuleInfo.name) {
             if (cb) {
@@ -128,10 +115,13 @@ module.exports = class MenuStack {
 
         const loadOpts = {
             name: name,
-            client: self.client,
+            client: this.client,
         };
 
-        if (currentModuleInfo && currentModuleInfo.menuFlags.includes('forwardArgs')) {
+        if (
+            currentModuleInfo &&
+            currentModuleInfo.menuFlags.includes(MenuFlags.ForwardArgs)
+        ) {
             loadOpts.extraArgs = currentModuleInfo.extraArgs;
         } else {
             loadOpts.extraArgs = options.extraArgs || _.get(options, 'formData.value');
@@ -140,33 +130,16 @@ module.exports = class MenuStack {
 
         loadMenu(loadOpts, (err, modInst) => {
             if (err) {
-                //  :TODO: probably should just require a cb...
-                const errCb = cb || self.client.defaultHandlerMissingMod();
+                const errCb = cb || this.client.defaultHandlerMissingMod();
                 errCb(err);
             } else {
-                self.client.log.debug({ menuName: name }, 'Goto menu module');
+                this.client.log.debug({ menuName: name }, 'Goto menu module');
 
                 if (!this.client.acs.hasMenuModuleAccess(modInst)) {
                     if (cb) {
                         return cb(Errors.AccessDenied('No access to this menu'));
                     }
                     return;
-                }
-
-                //
-                //  Handle deprecated 'options' block by merging to config and warning user.
-                //  :TODO: Remove in 0.0.10+
-                //
-                if (modInst.menuConfig.options) {
-                    self.client.log.warn(
-                        { options: modInst.menuConfig.options },
-                        'Use of "options" is deprecated. Move relevant members to "config" block! Support will be fully removed in future versions'
-                    );
-                    Object.assign(
-                        modInst.menuConfig.config || {},
-                        modInst.menuConfig.options
-                    );
-                    delete modInst.menuConfig.options;
                 }
 
                 //
@@ -182,9 +155,9 @@ module.exports = class MenuStack {
                     //  in code we can ask to merge in
                     if (
                         Array.isArray(options.menuFlags) &&
-                        options.menuFlags.includes('mergeFlags')
+                        options.menuFlags.includes(MenuFlags.MergeFlags)
                     ) {
-                        menuFlags = _.uniq(menuFlags.concat(options.menuFlags));
+                        menuFlags = [...new Set(options.menuFlags)]; // make unique
                     }
                 }
 
@@ -195,16 +168,12 @@ module.exports = class MenuStack {
 
                     currentModuleInfo.instance.leave();
 
-                    if (currentModuleInfo.menuFlags.includes('noHistory')) {
+                    if (currentModuleInfo.menuFlags.includes(MenuFlags.NoHistory)) {
                         this.pop();
-                    }
-
-                    if (menuFlags.includes('popParent')) {
-                        this.pop().instance.leave(); //  leave & remove current
                     }
                 }
 
-                self.push({
+                this.push({
                     name: name,
                     instance: modInst,
                     extraArgs: loadOpts.extraArgs,
@@ -216,17 +185,19 @@ module.exports = class MenuStack {
                     modInst.restoreSavedState(options.savedState);
                 }
 
-                const stackEntries = self.stack.map(stackEntry => {
-                    let name = stackEntry.name;
-                    if (stackEntry.instance.menuConfig.config.menuFlags.length > 0) {
-                        name += ` (${stackEntry.instance.menuConfig.config.menuFlags.join(
-                            ', '
-                        )})`;
-                    }
-                    return name;
-                });
+                if (this.client.log.level() <= bunyan.TRACE) {
+                    const stackEntries = this.stack.map(stackEntry => {
+                        let name = stackEntry.name;
+                        if (stackEntry.instance.menuConfig.config.menuFlags.length > 0) {
+                            name += ` (${stackEntry.instance.menuConfig.config.menuFlags.join(
+                                ', '
+                            )})`;
+                        }
+                        return name;
+                    });
 
-                self.client.log.trace({ stack: stackEntries }, 'Updated menu stack');
+                    this.client.log.trace({ stack: stackEntries }, 'Updated menu stack');
+                }
 
                 modInst.enter();