+++ WHATSNEW.md
@@ -16,6 +16,7 @@ This document attempts to track **major** changes and additions in ENiGMA½. For
 * CombatNet has shut down, so the module (`combatnet.js`) has been removed.
 * New `NewUserPrePersist` system event available to developers to 'hook' account creation and add their own properties/etc.
 * The signature for `viewValidationListener`'s callback has changed: It is now `(err, newFocusId)`. To ignore a validation error, implementors can simply call the callback with a `null` error, else they should forward it on.
+* The Menu Flag `popParent` has been removed and `noHistory` has been updated to work as expected. In general things should "Just Work", but do see [UPGRADE](UPGRADE.md)!
 
 ## 0.0.13-beta
 * **Note for contributors**: ENiGMA has switched to [Prettier](https://prettier.io) for formatting/style. Please see [CONTRIBUTING](CONTRIBUTING.md) and the Prettier website for more information.